// src/app/store/models/cap.model.ts
export interface Cap {
    Personal: number;
    Vehicle: number;
    Quiz: number;
    BrokerManual: number;
    BrokerManualAcknowledgement: number;
    DirectDeposit: number;
    WsibContract: number;
    ConsentForm: number;
    Aoda: number;
    Interview: number;
    Payment:number;
    AodaTraining: number;
    Total: number;
}